CHANGE-206 fix admin profiles code to also manage product types
2 problems not addressed originally.

zen_get_admin_pages was adding the product types entries to the admin menu. fixed by only getting the producvt type entries if we are not rendering menus
e.g. rendering the admin profiles page.

zen_get_menu_titles did not have an an entru for product types, so while each product type would render on the profiles page, it would not have a title.
manually added the _productTypes entry set to BOX_HEADING_PRODUCT_TYPES.
This should really test first to see if there are any product types, but will leave that for later