CHANGE-283 Allow for a custom count query in split page results class
The split page result classes parse the original query and rebuild it to attempt to add a count(*) in the query.
This repurposed query is used to provide a count of all the results to use in pagination displays.

The problem is that the parsing logic fails with some more complex queries. e.g. sub selects using GROUP BY clauses, and others

These changes to the catalog class allow for the use of a separate query being passed to the class that will be used to provide the count operation.
The change is backward compatible with previous versions.

The admin class is not amenable to the same change, so a copy of the catalog class has been provided to allow developers/contributors to use if necessary.