CHANGE-341 Updates to observer/notifier code to better support legacy procudural code
The observer/notifier system when used with purely OOP code allows the dynamic changing of observed class variables,
however when used with procedural code, it only allow the reading of parameters passed as an array in the notifier.

Changes allow the passing of up to six variables from the procedural context that can be amended in the observer class.

example changes have been made to the main_template_vars.php file within this commit.
Note, for backward compatibility, the first parameter passed to notify after the notifier id was traditionally an array,
and this can still be used to pass read only variables.
However if you only want to pass variables that can be amended, this first parameter should be passed as NULL.

e.g.
$zco_notifier->notify('NOTIFY_HEADER_INDEX_MAIN_TEMPLATE_VARS_PAGE_BODY', NULL, $tpl_page_body);

Furthermore, rather than relying on the update method of the observer, developers can now create methods that individually target a specific event id
e.g. based on the notifier above you could create a method in the observer class called
function updateNotifyHeaderIndexMainTemplateVarsPageBody(...

Note the event id is camelcased and prefixed with 'update'

This is useful when a single observer is servicing multiple event id's

CHANGE-341 oops, forgot to remove debugging code