CHANGE-367 Dashboard Widgets
Some calls to create the dashboard widgets objects were not passing the widgetInfo array. This meant custom templates would not load correctly.
Now just pass the widget key, and widgetInfo if available,
I guess some more refactoring is needed here as if widgetInfo is not availbale we should build it in the object.
Also add ajax/javascript code to trigger resetting of timer updates, when the settings for is submitted.