Request Class
See #65

This removes the old zcRequest class and replaces it with a namespaced Request class/
This class uses aura.web as its base. The reason for using aura.web is to allow for further expansion in the future,
and to address @lhungil requirement to have more control over the request/response

The commit simplifies the original zcRequest class, as that class was bloated and didn't really achieve the goals required.
Those goals were to enforce sanitization of every get/post parameter.
e.g. unless a sanitization had been set for a parameter, then that parameter would not be available through the request class,
and furthermore would not be available from the php superglobals.

The structure of how that enforced sanitization works needs some further thought.

So as a first step what we have is basically a facade around aura.web with some extra functionality
e.g. the ::has ::all methods

This allows us to update legacy code to use the new class, and later add on code to enforce sanitization.

As an adjunct, the aura.web class has been added as a git submodule. I'm still not sure if this is the best way to go as it requires
an extra step from a normal git fetch/pull
e.g. git submodule update --init --recursive
and i'm not sure what the gets us, rather than using composer. given that composer is only ever going to be used
by developers and final release code will be built manually.