Display first comment in right info box
At the moment Zen Cart is just displaying 'Comments' in the right info box when selecting orders in order list view, even if it's just some comments added by status change from a payment module - or a admin. This doesn't make any sense...

I believe it's better to just display this when there is a comment from the customer (from the first available order status), and at the same time display the comment so the admin can see if it's anything important without having to open every order first. This change in code will fix that.

At the same time I added a missing translation for ordered products bellow.