Modifying proposed display of login expiry notices
After more testing it seems like if "PA-DSS Admin Session Timeout Enforced? = On" 15 minutes will be the maximum, whatever is defined in "Admin Session Time Out in Seconds", if it's higher.
If defined value is same or lower than 900 it will be used.

Eg. if "PA-DSS Admin Session Timeout Enforced? = On" and "Admin Session Time Out in Seconds ="

    1800 it will "fallback" to, and timeout after, 900 (15 min)
    300 it will timeout after 5 minutes

Right? Then we need to change a bit in the code, into something like this.