Catalog Listing boxes/paginator refactoring
This is a refactoring of the current listingbox/paginator/querybuilder classes that are in v160

The refactoring is meant to make these classes more consistemt with PSR values, more easily testable. and easier to be overriden by 3rd party code.
The code has also been moved out of the classes directory and into the namespaced autoloaded library directory.

To simplify the code and make it more testable, the listingbox classes now take advantage of AuraPHP Dependency Injection, although at the moment we are pretty much using this
as a service locator, rather than a proper DI infrastructure. (mainly because we still have a lot of procedural code to support)

I am currently working on documentantion for these classes, based on our new codex server. his should be up and running soon.

In terms of unittesting there is good coverage for the basic classes, but lots more needs to be done here. At least we have a a structure to add unit tests for any bugs found.

There is also lots of scope to add more notifiers.