Poor scoping in action class means $zcRequest is not available in templates
See Issue #200

tplHeader directly addresses the $zcRequest variable when multiple languages are installed.
However on the home page which is served via an Action Class, $zcRequest is not in the global scope, as templates are rendered using a method of the ActionAdminBase class.

Need to add a template variable to pass in the current cmd param so this can be rendered correctly in the tplHeader template.