Add "hostname" override for email generator.
Although rare, some heavily "shared" servers are configured to handle sub-sub-domains, and the auto-detection of the email hostname isn't always accurate, especially if SERVER_NAME isn't passed correctly to PHP.

This change allows the ability to specify the hostname to use in Message-Id and Received headers and as default HELO string.

To use: define EMAIL_HOSTNAME in the extra_configures or extra_datafiles folder on both catalog and admin sides.