Merge #292 Send 503 header on db error message
When a database error occurs, we output a generic message, but it still sends a normal 200 response header, which doesn't tell search engines to skip it.

This sends the 503-service-unavailable header to tell SE's to come back later instead of indexing the error message.

(We were already sending the 503 header, but only when also redirecting to nddbc.html when a "not connected" condition was found ... which is too narrow criteria.)