Initial commit for refactored pagination classes
This is a refactoring of the previous new paginator classes, which were meant as a replacement for
the old split_page_results classes in admin/catalog.
The classes have been simplified and moved into the autoloaded library directory.
Note also, that I have not removed the old paginator classes, as these are still used by the listingBox
infrastructure. Once that is refactored they will be removed.