Cleanup of autoloading in admin/catalog
Previously Admin had it's own autoloader and includes/library directory.
This meant we would have had to duplicate code and directory structures in catalog to support autoloading there.
This commit moves the canonical library directory to catalog, and uses an autoload_namespaces.php file in in the catalog includes directory
for both admin and catalog.