Refactoring of catalog listingbox code
This refactors code into the library autoloader directory, as well as tidying up the object graph.
There's still some work to do here, mainly to do do with the use of GLOBALS/SESSION stuff.
While there is a a lot of code coverage for unittests, this doesn't really cover off real world usage, so that needs addressing as well.

Also need to update the documentation to match the new api/object graph.