Add new issetorArray function
Some upcoming code in v1.6 needs to test frequently git commit -a  whether an array key exists, and if not set a default value within a ternary structure.
(and we do this is legacy code as well)

As an example

Say we have an array called $somearray, and we want to test that $somearray['key'] exists, and if it doesn't set a default value.

e.g.
if (isset($somearray['key']) {
  return  $somearray['key']
 } else {
  return 'default';
 }
or as a ternary

return (isset($somearray['key']) ? $somearray['key'] ? 'default';

In some cases the above, even when used with a ternary can require a line that extends over the 80 character line limit

using the function rewrites the above as

return issetorArray($somearray, $key, 'default')

reducing the code footprint.

It should be noted that this function mimics the upcoming php7 Null Coalesce Operator https://wiki.php.net/rfc/isset_ternary#vote