Admin  LEAD Controllers
Please also see https://github.com/zencart/zc-v1-series/issues/63

LEAD stands for
 List
 Edit
 Add
 Delete

and is a way of defining (via simple array structures) how to build a page that manages a database table(and possible related tables).

The Admin pages covered by this commit are

  countries
  zones
  geo_zones
  tax_classes
  tax_rates

  currencies
  lanuages
  orders_status

  record_artists
  record_companies

although more will be added in the future.

The structure of the classes used and notifiers defined also is meant to allow for 3rd party contributions to add functionality to the pages without touching core code,
however there probably needs to be more work done here. Not only on notifier points, but alos the way derivedItems and field formatter callabkes work.