Change how some base classes are autoloaded.
Currently we preload a lot of class files via the confi.core.php autoload system.
Despite it's name the current autoload system does not lazy load those classes.
This means in some cases, that we are loading classes that are not needed, and are therefore wasting memory.

Have moved a lot of those to use the Aura Autoloder setClassFiles method.