zcInstall pattern matching when reading config file entries was broken.
The problem was that the config file would be loaded as a contigous string, and the regex was trying to anchor at the start of the string.
There is probably a better way to do this but for now, will load the config file as an array of lines then implode with a custom seprarator.
Can then key the regex to look for the separator.