"shared" template tier not worikng for sideboxes
While testing the shared feature on side boxes I noticed that the files in include/modules/sideboxes/shared were ignored. I think this is because there is a condition missing in the column_left.php, column_right.php, and column_single.php.
After adding the conditions to the files the php files in the includes/modules/sideboxes/shared folder are used, when present.