Maintain order status that is above default on PayPal batch capture
Discussion held back on ZC 1.5.1 at https://www.zen-cart.com/showthread.php?214423-Order-status-changes-after-batch-capture-on-PayPal-website where store owner observed that the status of orders that were previously manually updated inside of the ZC admin were lowered once batch capture was performed on PayPal site.

The change proposed above was reported as working at least before the and was added to the added if statement.  And statement intent is to apply/allow the old status to remain if and only if the transaction is one that isn't intended to reduce the status.  For example, refunding, reversing, denying or failing a transaction should not allow the status to remain elevated through this process, but instead take the status intended for those conditions.