Eliminate potential boolean result
Per discussion in merge request thread with main application, the parentheses around the sql assignment could be programatically incorrectly considered an evaluation for true/false where instead, the expectation is to evaluate to a string.

Honestly don't recall why a set of parentheses was used to begin with other than perhaps there was some initial boolean evaluation being performed when attempting to answer the user's question that was not fully removed at time of posting.