Test Option_name if no option_values assigned
If there are no option_values assigned, then loop through the option_values that exist.  If the option_names expect to have an option_value then
perform the action that indicates to user that option_values should exist.  Intention is to use the function zen_option_name_no_values that is in:
includes/functions/functions_lookups.php (added per this branch).  This is unfortunately untested and therefore the function may need to be
in the admin/includes/functions/general.php file instead.

Added function to the admin/includes/functions/general.php file after cleanup.