Appropriately applied new foreach sql_result array
Corrected an issue with $opt_type in line 2166 that was a unit of a sql result but did not reference the single query result as an array, but instead had
just used the variable for comparison.  Upon further review of the code changes to implement this methodology, decided better to incorporate the
array key into the above code.