Corrected application of foreach on queryfactoryResult in includes/functions/functions_lookups.php
Corrected an issue with $opt_type in line 274 that was a unit of a sql result but did not reference the single query result as an array, but instead had
just used the variable for comparison.  Upon further review of the code changes to implement this methodology, decided better to incorporate the
array key into the above code.