Admin Side remove zen_db_output usage Issue #270
Expanding the data for zen_db_output uses on the admin side in preparation
to 1) remove all usage of zen_db_output, 2) redefine zen_output_string_protected
so that it functions on the admin side the same as on the store side.
Requires substituting function into an equivalent standard function.