Fix bindvars to honor string entries that contain 'null'
Because of the way bindvars worked some form entries would
assume that if the entry contained the string 'null' that the whole entry
should be set as a mysql null.
e.g. If your surname was NullSurname then the bindvars would insert a mysql Null
rather than the proper string