Zc160 (#1)
* does not redirect on invalid product_id

* Remove block of dead code

* Maintain order status that is above default on PayPal batch capture

Discussion held back on ZC 1.5.1 at https://www.zen-cart.com/showthread.php?214423-Order-status-changes-after-batch-capture-on-PayPal-website where store owner observed that the status of orders that were previously manually updated inside of the ZC admin were lowered once batch capture was performed on PayPal site.

The change proposed above was reported as working at least before the and was added to the added if statement.  And statement intent is to apply/allow the old status to remain if and only if the transaction is one that isn't intended to reduce the status.  For example, refunding, reversing, denying or failing a transaction should not allow the status to remain elevated through this process, but instead take the status intended for those conditions.

* Eliminate potential boolean result

Per discussion in merge request thread with main application, the parentheses around the sql assignment could be programatically incorrectly considered an evaluation for true/false where instead, the expectation is to evaluate to a string.

Honestly don't recall why a set of parentheses was used to begin with other than perhaps there was some initial boolean evaluation being performed when attempting to answer the user's question that was not fully removed at time of posting.

* Add $new_status new status debug notification when changed.

* Remove added/extra ipn_debug_email notifications

In prepping the file for incorporation, added additional notification of the change in $new_status. Identified however that only one total notification needed, so it is the only one that remains from this modification and it is found inside an if statement.

* Cleaning up dead code, making brackets match

* Fixes 915

* Add back some Travis efficiencies

* Update repo name in readme and composer

* Update dependencies to more modern versions

* Remove selenium components until webtests are rebuilt

* Add multiple notifiers to the Admin for prod/cat/gv/sale/special/ezpage/currency/orders/etc

Closes #139

* Update repo name

* fix checkout bug caused by use of $$

* attempt to fix travis PHP7 errors

* attempt to fix travis PHP7 errors