Fix error-log on widget-install (#1133)
Given that the process of adding a selected widget to the dashboard doesn't require rendering that widget until the page is ready, calling the template isn't necessary. And, since the prepared vars from the widget class haven't been loaded, the vars aren't ready for the template and thus an error is logged.  Alternate fix would be to run the `loadClass` method, but that appears to be wasted CPU activity, hence these lines are simply disabled at the moment.