On update in-cart quantity remains if qty still correct
Fixes #1163

In an update performed to notify customers that the quantity in the cart had been modified (primarily addressing when adding product to the cart), the same code had been applied to the actionUpdateProduct function which if before the refresh the quantity in the cart plus the quantity to which it is to be set is greater than the available stock, then the quantity to be presented would be modified.

IE: store is setup to not allow going "negative" (allow checkout if out-of-stock is false), if a product has 10 in stock, 9 are added to the cart, the cart is then visited, and while on the cart if the refresh button is pressed, the quantity of that product would become 1. 

The above change removes that code but maintains notification to the customer if the quantity were adjusted for reasons such as attempting to add more than what is in stock or more than the maximum allowable or in other words if the quantity has been modified to be the quantity that currently existed in the cart or the maximum allowed.