Removing View code from controllers
Currently a lot of controllers admin/lead/ajax etc directly affect view code, including tplVars.
This change creates a first class view object (via a factory) that also encapsulates a Template Vars manager.
The idea is to remove the direct use of tplvars within controllers and related classes in favour of the
View -> tplVarManager code.
This commit is only a first stage towards that.