mc12345678 - Implement PayPal In-Context against the Express Button.
Added jscript_ files for the following pages where a customer could
land as a part of initiating paymnt and continuing through the
payment process:
shopping_cart,
login,
time_out,
checkout_shipping,
checkout_payment,
checkout_confirmation,
checkout_success,
index

Added a payment/paypal file that is referenced by the jscript_ files
to provide a consistent look at all uses.
This also required a change to the tpl_ec_button file because of the
way that paypal has implemented the display of their button. Now
if the information needed to display In-Context is entered and active
then the PayPal javascript will insert it's own image if it is
enabled.  If javascript is disabled, the ZC image will display.
This way there is always an image to click even if in context is
having an issue.

Added an auto_loader and observer to clear the paypal token session data
if a visitor has selected the cancel option from PayPal.  This also
required changing the process_button code of the payment file paypalwpp.php.
Also the checkout_confirmation templaate file was modified to include the
paypal data identified in the examples provided to initiate paypal express
in context.

Added code to the function process_button of
includes/modules/payment/paypalwpp.php to process the return function data
once paypal incontext is complete if payment is made through the
confirmation page instead of having previously paid through paypal
incontext earlier in the checkout process.  This was particularly necessary
when processing a failed payment such as using the default address
information of a demo product store where ZC provided the address by stepping
through checkout to the end using the demo data for address information.

Fixes #1192