Support table prefixes for 1.3.8 upgrade in zc_installer
Prior to ZC 1.5.5, the zc_installer included code to process the upgrade
script from ZC 1.3.8 to 1.3.9; however, in the redevelopment for ZC 1.5.5,
operations did not include processing SQL commands such as only DROP TABLE
though DROP TABLE IF EXISTS was supported.  This code was replaced and
expanded to support DROP TABLE and DROP TABLE IF EXISTS.

Also the code did not support the SQL command of RENAME TABLE.

To support the revised code the ZC 1.3.9 sql (upgrades 1.3.8 to 1.3.9)
required minor modification as well.

No other related/earlier SQL was identified as needing this change.