Special and SaleMaker with attributes sales correction
Priced-By-Attr, Price Factor, Special checkout correction

When viewing a product in the store that has been priced by attributes or not,
and it possibly has a price factor, special, or sale to be applied to the
attribute(s) the attribute information on the page has not previously
appeared correctly.

When adding a product to the cart that has been priced by attributes or not,
and it possibly has a price factor, special, or sale to be applied to the
attribute(s) the shopping_cart page appears to reflect all of the correct
prices; however, once the customer moves forward to checkout, the price
does not calculate correctly.

This appears to be a result of a number of factors that have either been
pulled out of some more core code to try to provide more of a
unique/editable solution or it was something that never made it quite into
the code area that would provide the desired functionality.

This commit seeks to correct two major issues, presentation of information
at the product info page and at checkout_payment and beyond.

This is heavily discussed and explained at this ZC post:
https://www.zen-cart.com/showthread.php?221940-attributes-price-factor-special-price-incorrect-totals&p=1326843#post1326843