Handle product priced-by-attribute with non-zero base product price
Adjust functions calculate and attributes_price to support the product
itself having a price in addition to the product being priced by attributes
and being able to apply a discount or other price reduction.

Both functions needed this modification to support not only the shopping_cart
display but transfer of the shopping_cart data to the checkout_ process.