Improve catalog side of zen_get_attributes_price_final
Now passing the status of the product being priced by attribute so that
it is not calculated within the function but instead outside of it.  Particularly
since this function is called by the attributes file, it is likely to
need to know many times that the product is or is not priced by attribute.
This way it is determined one time for the product and referenced multiple
times instead of asking the database several times and processing the
response for each occurrence.