Add function to identify the presence of a date span overlap
zen_datetime_overlap offers the ability to identify the overlap relationship
between two provided date ranges with consideration of the frame of
reference by which to make the comparison.  Ie. should dates be compared
with concern of their overlap in the future, in the past only or an
overlap existing at all.

The function was originally considered to support the addition of specials
to a product such that it would not be possible from the admin entry point
to insert a new special that does not overlap with an existing (active)
special.  At least initially until it is determined/decided that specials
can be applied to specials for example.

This is/was also discussed in the ZC forum at: https://www.zen-cart.com/showthread.php?222038-Can-there-be-more-than-one-quot-Special-quot-for-a-product