Updated notes for zen_get_discount_calc and zen_get_products_sale_discount_type.
While reworking portions of this code a few things helped to try to get it
right.  One was the notes about how the discount_type is determined provided
detailed calculations, but the data result does not match how ZC currently
calculates. Then there are the notes about when applying discounts how
one has arrived at that discount type so that it can be understood what
needs to happen moving forward.

In reviewing the various discount types discovered that there are tests
for some values that no longer actually function/operate because the value
is never reached.