Alternative ensurance that attributes have values
This is a modified version of the code proposal that fixes #1454.

The operation of the testing has been simplified to having only two returns instead of a series of logic "tests" of when false should be returned providing true if and only if a price `options_values_price` is identified as being non-zero, otherwise false.