Add define check
Add definition check to ATTRIBUTES_PRICE_FACTOR_FROM_SPECIAL use within functions_prices.php. As a result of chicken/egg loading, a definition for the function may not yet exist at the point of loading, but during execution it is expected to be defined prior to use. That definition could be in the database, an additional file loaded prior to the functions or in the execution of say the shopping_cart class.