V156 (#10)
* Alternative ensurance that attributes have values

This is a modified version of the code proposal that fixes #1454.

The operation of the testing has been simplified to having only two returns instead of a series of logic "tests" of when false should be returned providing true if and only if a price `options_values_price` is identified as being non-zero, otherwise false.

* Remove SHOW_GV_QUEUE define from english.php into ot_gv settings

Fixes #1420

* Expand comment text field for product-options from 64 to 255 characters

Ref: https://www.zen-cart.com/showthread.php?222625-Option-name-comment-hyperlink

* Remove unneeded AIM code for admin refunds page

* Add customer telephone number to order-confirmation email

* Update Square module to v0.82

* Update README

* Payment module updates for number formatting

* Some tidying in alpha prep stage

* Add map links to admin order page for manual address verification

* Add notifier hooks to currency-updater

* Currency exchange message output rounding to 8 decimals, to match db

* Coupon enhancements - reporting, exporting, additional restrictions

* Simplify TLD logic, since we're only operating under SSL

* Set admin "title" so browser tabs show current page name

* Add .gitignore/.htaccess to admin backups dir

This helps keep db backups out of version-control, and specifically denies unauthorized downloads.

* Add additional order fields to database, for easier order-reconstruction/edit

* Force log-exporter to only allow supported filenames

* Use store's defined DATE_FORMAT for activity-log warning message

Ref: https://www.zen-cart.com/showthread.php?222649-date-format-fix-suggestion-in-admin-activity-alert

* Include @lat9's Styleable Min/Max/Unit/Mixed Display

Ref: http://www.zen-cart.com/downloads.php?do=file&id=1922

* Update currency-exchange calc server connection parameters

* Restore no-default on tinytext fields - partial reversal of #1470

strict schema rules don't like a "default" value on text/blob fields, including tinytext

* Merge #1474 and update code deprecation comments

* Docs

* Add Sales-Report-With-Graphs plugin

* Update Square module to v0.90, fixing auto token refresh problem.

* Fix demo typos

Closes #1482

* Rename "in stock"/"out of stock" to "enabled"/"disabled"

Closes #1476

* Remove unused `final_price` field from customers_basket table.

Closes #1285

* Fixes #1457 (M'ple GC's purchased ...)

* Add language-change notifier hooks

* Compatibility changes to accommodate OPC plugin

* Abort ajax requests from spiders

Closes #1296

* Update PHPMailer to 5.2.24

* Docs

* Ignore sub-arrays in zen_get/post_all_get_params

Nested arrays for get/post values are not used internally in any way that needs re-echoing to the page. Skipping them to avoid throwing false errors when bots/probers do unauthorized stuff.
Ref https://www.zen-cart.com/showthread.php?222694

* Update embedded jQuery to latest 1.x: 1.12.4

* Docs

* Fix confusing text related to copying quantity discounts when cloning products

* Store Manager - reset master_categories_id to use original insert order

By bypassing indexes, it now resets to first-inserted master_categories_id which is more consistently the intended master.

* Fixes PR #1479 bypass index on store_manager reset for master_categories_id

* wrong/unnecessary line break

* typo in table name

product_options should be products_options

* Add missing <tr> brackets

Brackets for the row in this table were missing

* Update preview_info_meta_tags.php

* Update preview_info_meta_tags.php

* Update preview_info_meta_tags.php

* remove tell a friend remnants

For reference see:
https://www.zen-cart.com/showthread.php?222776-Remnants-of-obsolete-Tell-a-Friend-feature-in-ZC-1-5-5x

* Update functions_email.php

* Revise ckeditor helper to detect class names more universally

* Allow admin javascript to have its own file instead of altering core template components

* Update jQuery/Bootstrap dist components

* Update FontAwesome components

* Split TEXT_VISTORS_CART to create TEXT_CART_HELP for shopping_cart page

Closes #1080

* Harmonize js versions

* Add function to identify the presence of a date span overlap

zen_datetime_overlap offers the ability to identify the overlap relationship
between two provided date ranges with consideration of the frame of
reference by which to make the comparison.  Ie. should dates be compared
with concern of their overlap in the future, in the past only or an
overlap existing at all.

The function was originally considered to support the addition of specials
to a product such that it would not be possible from the admin entry point
to insert a new special that does not overlap with an existing (active)
special.  At least initially until it is determined/decided that specials
can be applied to specials for example.

This is/was also discussed in the ZC forum at: https://www.zen-cart.com/showthread.php?222038-Can-there-be-more-than-one-quot-Special-quot-for-a-product

* Correct array operation when calling zen_option_name_base_expects_no_values

The way the code ended up being written was that if an array is presented
to the function then the code would error because the last internal
line of code would set the $sql2 to just a right parentheses instead
of adding a parenthesis at the end of the string.

* Mc12345678 remove zen_get_buy_now_qty from template (#13)

Update product main_template_vars to include the buy_now_qty and and template file to reference previously collected data

Adding the result of the zen_get_buy_now_qty into the product main_template_vars removes the processing (and potential modifications necessary to support alternate processes) from the template page.  This allows the template to present the information that is needed/expected without relying on additional processing or dependency on internal functions.

This has also been considered as a way to solve an issue that can occur when using the product option type Grid as offered by Products attributes grid where specifically the use of units or minimum quantities are intended to only impact the individual attribute selections not the overall "product".

* Updates to zc_install

Fixes some strict type errors
Fixes typo in dist-configure

* Refactor update_product for strict data types

* Support ISO-8601 style date-of-birth date (merge #1515) (#1518)

Support ISO-8601 style date-of-birth date

* Fix graph display when counter-history is empty

Closes #1526